Nova liberty-3 milestone (12.0.0.0b3)

+++ nova/api/openstack/compute/consoles.py
@@ -13,12 +13,16 @@
 #    License for the specific language governing permissions and limitations
 #    under the License.
 
-import webob
 from webob import exc
 
+from nova.api.openstack import extensions
 from nova.api.openstack import wsgi
 from nova.console import api as console_api
 from nova import exception
+
+
+ALIAS = 'os-consoles'
+authorize = extensions.os_compute_authorizer(ALIAS)
 
 
 def _translate_keys(cons):
@@ -30,9 +34,7 @@ def _translate_keys(cons):
 
 
 def _translate_detail_keys(cons):
-    """Coerces a console instance into proper dictionary format with
-    correctly mapped attributes.
-    """
+    """Coerces a console instance into proper dictionary format with detail."""
     pool = cons['pool']
     info = {'id': cons['id'],
             'console_type': pool['console_type'],
@@ -43,50 +45,84 @@ def _translate_detail_keys(cons):
     return dict(console=info)
 
 
-class Controller(object):
+class ConsolesController(wsgi.Controller):
     """The Consoles controller for the OpenStack API."""
 
     def __init__(self):
         self.console_api = console_api.API()
 
+    @extensions.expected_errors(())
     def index(self, req, server_id):
         """Returns a list of consoles for this instance."""
+        context = req.environ['nova.context']
+        authorize(context, action='index')
+
         consoles = self.console_api.get_consoles(
-                                    req.environ['nova.context'],
-                                    server_id)
+                req.environ['nova.context'], server_id)
         return dict(consoles=[_translate_keys(console)
                               for console in consoles])
 
+    # NOTE(gmann): Here should be 201 instead of 200 by v2.1
+    # +microversions because the console has been created
+    # completely when returning a response.
+    @extensions.expected_errors(404)
     def create(self, req, server_id, body):
         """Creates a new console."""
+        context = req.environ['nova.context']
+        authorize(context, action='create')
+
         try:
             self.console_api.create_console(
-                                req.environ['nova.context'],
-                                server_id)
+                req.environ['nova.context'], server_id)
         except exception.InstanceNotFound as e:
             raise exc.HTTPNotFound(explanation=e.format_message())
 
+    @extensions.expected_errors(404)
     def show(self, req, server_id, id):
         """Shows in-depth information on a specific console."""
+        context = req.environ['nova.context']
+        authorize(context, action='show')
+
         try:
             console = self.console_api.get_console(
                                         req.environ['nova.context'],
                                         server_id,
                                         int(id))
-        except exception.NotFound as e:
+        except exception.ConsoleNotFound as e:
             raise exc.HTTPNotFound(explanation=e.format_message())
         return _translate_detail_keys(console)
 
+    @wsgi.response(202)
+    @extensions.expected_errors(404)
     def delete(self, req, server_id, id):
         """Deletes a console."""
+        context = req.environ['nova.context']
+        authorize(context, action='delete')
+
         try:
             self.console_api.delete_console(req.environ['nova.context'],
                                             server_id,
                                             int(id))
-        except exception.NotFound as e:
+        except exception.ConsoleNotFound as e:
             raise exc.HTTPNotFound(explanation=e.format_message())
-        return webob.Response(status_int=202)
 
 
-def create_resource():
-    return wsgi.Resource(Controller())
+class Consoles(extensions.V21APIExtensionBase):
+    """Consoles."""
+
+    name = "Consoles"
+    alias = ALIAS
+    version = 1
+
+    def get_resources(self):
+        parent = {'member_name': 'server',
+                  'collection_name': 'servers'}
+        resources = [
+            extensions.ResourceExtension(
+                'consoles', ConsolesController(), parent=parent,
+                member_name='console')]
+
+        return resources
+
+    def get_controller_extensions(self):
+        return []
